requirements.txt
anthropic==0.86.0
pdfplumber==0.11.9
python-dotenv==1.2.2
pydantic==2.12.5
httpx==0.28.1
flask>=3.0,<4.0
waitress>=3.0,<4.0
